Fix Sonar reported bugs
Improve code coverage

Issue-ID: SDNC-1603

+++ sdnr/wt/devicemanager-onap/onf14/provider/src/main/java/org/onap/ccsdk/features/sdnr/wt/devicemanager/onf14/impl/interfaces/Onf14WireInterfaceNotificationListener.java
@@ -95,7 +95,7 @@ public class Onf14WireInterfaceNotificationListener implements WireInterface20Li
         FaultlogEntity faultAlarm = new FaultlogBuilder().setObjectId(notification.getObjectIdRef().getValue())
                 .setProblem(notification.getProblem()).setTimestamp(notification.getTimestamp())
                 .setNodeId(this.netconfAccessor.getNodeId().getValue()).setSourceType(SourceType.Netconf)
-                .setSeverity(mapSeverity(notification.getSeverity())).setCounter(notification.getCounter().intValue())
+                .setSeverity(mapSeverity(notification.getSeverity())).setCounter(notification.getCounter())
                 .build();
         serviceProvider.getFaultService().faultNotification(faultAlarm);
         serviceProvider.getWebsocketService().sendNotification( notification, netconfAccessor.getNodeId(),
